Change constant for trigger_error
According to http://php.net/manual/en/function.trigger-error.php
```
error_type
The designated error type for this error. It only works with the E_USER family of constants, and will default to E_USER_NOTICE.
```

So to discover what is happening in exceptions I've only change to E_USER family for compliance.